Refactor, move App.js to src/
